create animal

+++ 49 animals/index.js
@@ -1,6 +1,7 @@
 const express = require("express");
 const bodyParser = require("body-parser");
 const { v4: uuidv4 } = require("uuid");
+const animalZoo = require("./data/animals");
 const app = express();
 const port = 80;
 
@@ -8,30 +9,24 @@ app.use(express.static("public"));
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
 const fs = require("node:fs");
-const animalZoo = [
-  { name: "Lion", species: "Panthera leo", age: 10 },
-  { name: "Tiger", species: "Panthera tigris", age: 12 },
-  { name: "Elephant", species: "Loxodonta africana", age: 15 },
-  { name: "Giraffe", species: "Giraffa camelopardalis", age: 20 },
-  { name: "Zebra", species: "Equus zebra", age: 8 },
-  { name: "Hippopotamus", species: "Hippopotamus amphibius", age: 25 },
-  { name: "Crocodile", species: "Crocodylus niloticus", age: 30 },
-  { name: "Flamingo", species: "Phoenicopterus", age: 5 },
-  { name: "Penguin", species: "Aptenodytes forsteri", age: 3 },
-  { name: "Kangaroo", species: "Macropus", age: 7 },
-  { name: "Koala", species: "Phascolarctos cinereus", age: 4 },
-  { name: "Kookaburra", species: "Dacelo novaeguineae", age: 2 },
-  { name: "Platypus", species: "Ornithorhynchus anatinus", age: 6 },
-  { name: "Wallaby", species: "Macropus", age: 5 },
-  { name: "Dingo", species: "Canis lupus dingo", age: 8 },
-];
 
-fs.writeFileSync("./data/animals.json", JSON.stringify(animalZoo));
+const compareByName = (a, b) => {
+  const nameA = a.name.toLowerCase();
+  const nameB = b.name.toLowerCase();
+  if (nameA < nameB) return -1;
+  if (nameA > nameB) return 1;
+  return 0;
+};
+
+if (!fs.existsSync("./data/animals.json")) {
+  fs.writeFileSync("./data/animals.json", JSON.stringify(animalZoo));
+}
 
 app.get("/", (req, res) => {
   let html = fs.readFileSync("./data/index.html", "utf8");
+  let nav = fs.readFileSync("./data/nav.html", "utf8");
   data = fs.readFileSync("./data/animals.json", "utf8");
-  data = JSON.parse(data);
+  data = JSON.parse(data).sort(compareByName);
   const listItem = fs.readFileSync("./data/listItem.html", "utf8");
   let listItems = "";
   data.forEach((li) => {
@@ -43,9 +38,32 @@ app.get("/", (req, res) => {
       .replace("{{AGE}}", li.age);
     listItems += liHtml;
   });
-  html = html.replace("{{LI}}", listItems);
+  html = html.replace("{{LI}}", listItems).replace("{{NAV}}", nav);
 
   res.send(html);
+});
+
+app.get("/create", (req, res) => {
+  let html = fs.readFileSync("./data/create.html", "utf8");
+  let nav = fs.readFileSync("./data/nav.html", "utf8");
+  let data = fs.readFileSync("./data/animals.json", "utf8");
+  data = JSON.parse(data);
+  html = html.replace("{{NAV}}", nav);
+  res.send(html);
+});
+
+app.post("/store", (req, res) => {
+  const name = req.body.name;
+  const species = req.body.species;
+  const age = parseInt(req.body.age);
+  const id = uuidv4();
+
+  let data = fs.readFileSync("./data/animals.json", "utf8");
+  data = JSON.parse(data);
+  data.push({ id, name, species, age });
+  data = JSON.stringify(data);
+  fs.writeFileSync("./data/animals.json", data);
+  res.redirect(302, "http://animals.test/");
 });
 
 app.listen(port, () => {
